fix issue #221
This should not be a bug. When you save selection, you need to “Add
Variable” first, at this step, you should specify to create an
“integer” field instead of “Real” (by default).

Still, there is an issue of “multiple time period” default check radio
should be “false” since there is no need to add multiple time periods
for “sel/unsel” variable.  This fix is for this issue.