fix issue 234
This fix only guarantee at default window size, the 999 permutation
occupies the whole width of the screen. The reason that it is skinny is
because there are more bins for 999 permutation, and the width of the
screen is not big enough for these bins. Of course the width of each
bin can be calculated as screen_width/#bins, however,  wxWidgets only
supports integer width when drawing rectangle — which means the width
of bin is rounded as floor(screen_width/#bins). And this round caused
the whole plot looks not using  the full width of screen. You will see
the same issue when you use 1999 permutation. The user can always
resize the window to get a better rendered plot.