fix issue #340
This is a very rare and weird issue: the reason of failure dbf-csv
conversion is caused by the column “UNIV”. It contains empty strings,
and the OGR csv plugin can’t handle this and report an error. Ideally
this should be fixed on OGR side, however, for a quick fix, we have a
workaround in GeoDa code for now.