fix issue #347
It works as supposed to do:  you need a “numeric” column to store 1/0
values as “selected”/“unselected”. So, the error message is shown if
you select, for example, a “string” column. This old design has “Add
Variable” to allow user can specify a new column to store this
selection variable.

The current fix: “Add Variable” will default to an integer column;
Error message will be updated.

If we got time, we can get a much more intuitive design.