fix issue #364
This fix is proposed to remember the position of wxListCtrl and reset
the last position. Since the SetScrollPos doesn’t work on wxListCtrl,
EnsureVisible(item) has to be used; however, the position jumps a
little — a highlight to next item is used to give the user a hint where
it it.