update Enter vs Apply
in wxWidgets, each Enter event has to be bind with a ctrl. e.g. an
input box.

The Enter works in LOWESS param dialog is because the focus is on an
input text box. So, when you hit ENTER, the code will trigger Apply
function.

In this case, there are two input boxes (now are bind with ENTER).
Other ctrls (e.g. var choice) are binding with Apply function directly.
For example, if you select a variable, the correlogram will be updated.

But if you hit ENTER not on an input text box, it won’t trigger Apply.