fix issue #379
There is a special purpose to highlight HC60 after group HR60-90. See
issue #364 https://github.com/lixun910/geoda/issues/364.

However, to do shift selection, one needs two mouse hits on the
wxListCtrl. In this case, the first hit is missed, so the system uses
the beginning as the first hit.

This fix solves this issue.