fix issue #380
The weights function has been modified to support both GAL and GWT for
spatial diagnostics in regression (see issue #202).

However,  this change caused the weights information has NOT been
copied completely due to newly added variable.

So, in the case: G cluster map first and then LISA map, the
un-completed weights copy caused the visualization bug (described in
#380)

This and previous commits are proposed to fix this issue.