fix issue #448
it's caused by introducing the transparency and  multi-layers to draw
maps. Legends are still drawn using old code, so they are not impacted.

This fix will address this bug, also the export type is limited to png
only — option “bmp” will be removed since it doesn’t support alpha
chanel.