#652 "smarter" way to size the legend and map relative to each other to avoid the white space
The current solution will provide a smarter way to layout legend and
map side-by-side in a fixed relation (resize the map window will only
change the size of the printed map in PNG/BMP/SVG). Ps print-out will
be controlled by the default page size of ps setup, which is good for
publication purpose.