#1223 #1222
in the correlogram, with display statistics on, when the pointer hovers
over the statistics part, the status bar shows “mouse position = (0,0)”.

After removing a grouped variable in the
time editor and starting over, with a new grouping. So, all previous
ones were removed (with the right click remove). The first label in
the time labels is not time0 but the edited label from the previous
grouped variable is there, in my example 2005, and the next one
is time0. Something must not be cleared properly.