fix a crashes related to Connect/Export Datasource dialog, includes: CreateGrid, MergeTable, etc.
Due to a wxwidgets 3.1 bug https://trac.wxwidgets.org/ticket/17459, I
have implemented our own version Connect/Export dialog which is not
based on wxDialog but wxFrame, which caused problems when open/close it
several times — e.g. the operations in CreateGrid and MergeTable.

Now, there is a temporal fix for this wx bug, so this fix is applying
the patch to restore the wxDialog based Connect/Export Datasource
dialog, and therefore fix bugs related to CreateGrid and MergeTable.