#1664
Yes, the "hovering to highlight” feature in connectivity map is very sensitive for now.  If one of the connectivity map window is activated by accident (e.g. mouse over to a white space),  all other selected/highlighted objects in other window will be removed as a result of linking from the connectivity map window. This should be fixed: maybe not doing anything when hovering to a white space, until one actually clicks on white space.