#1685 save as image quality
keep fixed size of legend in new save_as image feature. Note: this
legend size is relevant to the size of the map when save as image,
which has high resolution (300dpi).  E.g. you will see a “bigger” size
of legend in a 1600x800 pixel image than in a resized 800x400 pixel
image. However, the legend size in both images are the “same” relative
to the map beside.

add work around solution to deal with existing wxwidgets bug