update logic when saving a DBF file
When saving a DBF file, it is possible a .cpg file is created if a specific encoding (e.g. GB2312 for Chinese characters) is used in GeoDa Table. The encoding information could be loaded from original dataset or specified manually by a user using Table->Encode menu.

For development, the logic to handle additional files when saving a DBF file:
> 1. don't create a .prj file
> 2. don't create a .cpg file if no specific encoding is used in GeoDa
> 3. only create a .cpg file when a specific encoding is used in GeoDa