enhancement: let user to config if creating csvt file when saving csv file
CSVT file is to used to describe the type of each column of the layer and its optional width and precision in CSV file. It causes many issues in GeoDa, especially when original CSV file has been modified in other apps (e.g. Excel).

This fix will add a perference option "Create CSVT file when saving CSV", and it is set to "false" by default.